:zap: fix for post, setup for put
