+++ packages/nodes-base/nodes/HttpRequest/test/node/HttpRequest.get.test.ts
@@ -59,35 +59,3 @@ describe('Test HTTP Request Node, GET', () => {
 		test(testData.description, async () => equalityTest(testData, nodeTypes));
 	}
 });
-
-describe('Test HTTP Request Node, POST', () => {
-	const workflows = ['nodes/HttpRequest/test/node/workflow.post.json'];
-	const tests = workflowToTests(workflows);
-
-	beforeAll(() => {
-		nock.disableNetConnect();
-
-		nock('https://dummyjson.com')
-			.post('/todos/add', {
-				todo: 'Use DummyJSON in the project',
-				completed: false,
-				userId: 5,
-			})
-			.reply(200, {
-				id: 151,
-				todo: 'Use DummyJSON in the project',
-				completed: false,
-				userId: '5',
-			});
-	});
-
-	afterAll(() => {
-		nock.restore();
-	});
-
-	const nodeTypes = setup(tests);
-
-	for (const testData of tests) {
-		test(testData.description, async () => equalityTest(testData, nodeTypes));
-	}
-});
